add jpeg endpoint
* rename qtrender_png.py to qtrender_image.py
* wrap most of the functions in qtrender_image.py in QtImageRenderer class
* add support for RGB image format (JPEG doesn't work well with ARGB)
* add appropriate resource, renderer and url mapping
